+++ migrations/sqlite/000004_wiki_source_revision.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE knowledge_bases DROP COLUMN wiki_source_revision;